+++ src/android/src/main/java/com/reactnativecommunity/slider/drawables/DrawListenerRegistry.java
@@ -1,4 +1,4 @@
-package com.reactnativecommunity.slider.ReactSliderDrawable;
+package com.reactnativecommunity.slider.drawables;
 
 import android.view.View;
 import android.view.ViewTreeObserver;
